+++ fuel-derive/src/compression/compress.rs
@@ -58,18 +58,12 @@ fn construct_compressed(
     let bound_fields: TokenStream2 = variant
         .bindings()
         .iter()
+        .filter(|binding| !matches!(FieldAttrs::parse(&binding.ast().attrs), FieldAttrs::Skip))
         .map(|binding| {
-            let attrs = FieldAttrs::parse(&binding.ast().attrs);
             let ty = &binding.ast().ty;
             let cname = format_ident!("{}_c", binding.binding);
-
-            match attrs {
-                FieldAttrs::Skip => quote! {},
-                FieldAttrs::Normal => {
-                    quote! {
-                        let #cname = <#ty as ::fuel_compression::CompressibleBy<_>>::compress_with(&#binding, ctx).await?;
-                    }
-                }
+            quote! {
+                let #cname = <#ty as ::fuel_compression::CompressibleBy<_>>::compress_with(&#binding, ctx).await?;
             }
         })
         .collect();
@@ -77,11 +71,10 @@ fn construct_compressed(
     let construct_fields: TokenStream2 = variant
         .bindings()
         .iter()
+        .filter(|binding| {
+            !matches!(FieldAttrs::parse(&binding.ast().attrs), FieldAttrs::Skip)
+        })
         .map(|binding| {
-            let attrs = FieldAttrs::parse(&binding.ast().attrs);
-            if matches!(attrs, FieldAttrs::Skip) {
-                return quote! {};
-            }
             let cname = format_ident!("{}_c", binding.binding);
             if let Some(fname) = &binding.ast().ident {
                 quote! { #fname: #cname, }
@@ -103,40 +96,7 @@ fn construct_compressed(
     }
 }
 
-/// Generate a match arm for each variant of the compressed structure
-/// using the given function to generate the pattern body.
-pub fn each_variant_compressed<
-    F: FnMut(&synstructure::VariantInfo<'_>) -> TokenStream2,
->(
-    s: &synstructure::Structure,
-    compressed_name: &TokenStream2,
-    mut f: F,
-) -> TokenStream2 {
-    s.variants()
-        .iter()
-        .map(|variant| {
-            // Modify the binding pattern to match the compressed variant
-            let mut v2 = variant.clone();
-            v2.filter(|field| {
-                let attrs = FieldAttrs::parse(&field.ast().attrs);
-                !matches!(attrs, FieldAttrs::Skip)
-            });
-            v2.bindings_mut().iter_mut().for_each(|binding| {
-                binding.style = synstructure::BindStyle::Move;
-            });
-            let mut p = v2.pat().into_iter();
-            let _ = p.next().expect("pattern always begins with an identifier");
-            let p = quote! { #compressed_name #(#p)* };
-
-            let decompressed = f(variant);
-            quote! {
-                #p => { #decompressed }
-            }
-        })
-        .collect()
-}
-
-/// Derives `Compress` trait for the given `struct` or `enum`.
+/// Derives `Compressible` and `CompressibleBy` traits for the given `struct` or `enum`.
 pub fn derive(mut s: synstructure::Structure) -> TokenStream2 {
     s.add_bounds(synstructure::AddBounds::None)
         .underscore_const(true);
@@ -188,22 +148,6 @@ pub fn derive(mut s: synstructure::Structure) -> TokenStream2 {
         &mut w_impl_field_bounds_compress,
         syn::parse_quote! { Ctx: ::fuel_compression::ContextError },
     );
-
-    let mut w_impl_field_bounds_decompress = w_impl.clone();
-    for variant in s.variants() {
-        for field in variant.ast().fields.iter() {
-            let ty = &field.ty;
-            match FieldAttrs::parse(&field.attrs) {
-                FieldAttrs::Skip => {}
-                FieldAttrs::Normal => {
-                    where_clause_push(
-                        &mut w_impl_field_bounds_decompress,
-                        syn::parse_quote! { #ty: ::fuel_compression::DecompressibleBy<Ctx> },
-                    );
-                }
-            }
-        }
-    }
 
     let def = match &s.ast().data {
         syn::Data::Struct(v) => {
