+++ fuel-derive/src/compression/decompress.rs
@@ -6,13 +6,42 @@ use quote::{
 
 use crate::helpers::where_clause_push;
 
-use super::{
-    attribute::{
-        FieldAttrs,
-        StructureAttrs,
-    },
-    compress::each_variant_compressed,
+use super::attribute::{
+    FieldAttrs,
+    StructureAttrs,
 };
+
+/// Generate a match arm for each variant of the compressed structure
+/// using the given function to generate the pattern body.
+fn each_variant_compressed<F: FnMut(&synstructure::VariantInfo<'_>) -> TokenStream2>(
+    s: &synstructure::Structure,
+    compressed_name: &TokenStream2,
+    mut f: F,
+) -> TokenStream2 {
+    s.variants()
+        .iter()
+        .map(|variant| {
+            // Modify the binding pattern to match the compressed variant
+            let mut v2 = variant.clone();
+            v2.filter(|field| {
+                let attrs = FieldAttrs::parse(&field.ast().attrs);
+                !matches!(attrs, FieldAttrs::Skip)
+            });
+            v2.bindings_mut().iter_mut().for_each(|binding| {
+                binding.style = synstructure::BindStyle::Move;
+            });
+            let mut p = v2.pat().into_iter();
+            let _ = p.next().expect("pattern always begins with an identifier");
+            let p = quote! { #compressed_name #(#p)* };
+
+            let decompressed = f(variant);
+            quote! {
+                #p => { #decompressed }
+            }
+        })
+        .collect()
+}
+
 /// Construct original version of the struct from the compressed one
 fn construct_decompress(
     // The original structure to construct, i.e. struct name or enum variant path
@@ -23,13 +52,12 @@ fn construct_decompress(
         .bindings()
         .iter()
         .map(|binding| {
-            let attrs = FieldAttrs::parse(&binding.ast().attrs);
             let ty = &binding.ast().ty;
             let cname = format_ident!("{}_c", binding.binding);
 
-            match attrs {
+            match FieldAttrs::parse(&binding.ast().attrs) {
                 FieldAttrs::Skip => quote! {
-                    let #cname = Default::default();
+                    let #cname = ::core::default::Default::default(),
                 },
                 FieldAttrs::Normal => {
                     quote! {
@@ -64,7 +92,8 @@ fn construct_decompress(
         #original #construct_fields
     }
 }
-/// Derives `Decompress` trait for the given `struct` or `enum`.
+
+/// Derives `DecompressibleBy` trait for the given `struct` or `enum`.
 pub fn derive(mut s: synstructure::Structure) -> TokenStream2 {
     s.add_bounds(synstructure::AddBounds::None)
         .underscore_const(true);
