+++ fuel-derive/src/lib.rs
@@ -39,11 +39,11 @@ mod compression {
 
 synstructure::decl_derive!(
     [Compress, attributes(compress)] =>
-    /// Derives `Compressed` trait for the given `struct` or `enum`.
+    /// Derives `Compressible` and `CompressibleBy` traits for the given `struct` or `enum`.
     compression::compress::derive
 );
 synstructure::decl_derive!(
     [Decompress, attributes(compress)] =>
-    /// Derives `Compressed` trait for the given `struct` or `enum`.
+    /// Derives `DecompressibleBy` trait for the given `struct` or `enum`.
     compression::decompress::derive
 );
